+++ SmartGuardPortalv1/Controllers/Web/Features/MemoryController.cs
@@ -22,7 +22,7 @@ namespace SmartGuardPortalv1.Controllers
         [Authorize(Roles = "User")]
         public ActionResult Index()
         {
-            return View(db.Memories.Where(i => i.MemoryType == 0 || i.MemoryType == 2).ToList());
+            return View(db.Memories.Where(i => i.fkUserId == WebMatrix.WebData.WebSecurity.CurrentUserId).Where(i => i.MemoryType == 0 || i.MemoryType == 2).ToList());
         }
 
         //
@@ -37,8 +37,10 @@ namespace SmartGuardPortalv1.Controllers
                 {
                     return HttpNotFound();
                 }
-                return View(memory);
-           
+                else if(memory.fkUserId == WebMatrix.WebData.WebSecurity.CurrentUserId)
+                    return View(memory);
+                else
+                    return HttpNotFound();
             
            
                    
@@ -152,9 +154,10 @@ namespace SmartGuardPortalv1.Controllers
                 {
                     return HttpNotFound();
                 }
-                
-                return View(memory);
-                
+                else if(memory.fkUserId == WebMatrix.WebData.WebSecurity.CurrentUserId)
+                    return View(memory);
+                else
+                    return HttpNotFound();
                     
             
         }
@@ -191,8 +194,10 @@ namespace SmartGuardPortalv1.Controllers
                 {
                     return HttpNotFound();
                 }
-
-                return View(memory);
+                else if(memory.fkUserId == WebMatrix.WebData.WebSecurity.CurrentUserId)
+                    return View(memory);
+                else
+                    return HttpNotFound();
 
             
             
@@ -206,8 +211,11 @@ namespace SmartGuardPortalv1.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             Memory memory = db.Memories.Find(id);
-            db.Memories.Remove(memory);
-            db.SaveChanges();
+            if (memory.fkUserId == WebMatrix.WebData.WebSecurity.CurrentUserId)
+            {
+                db.Memories.Remove(memory);
+                db.SaveChanges();
+            }
             return RedirectToAction("Index");
         }
